// components/textarea/textAreaTargets.enum.ts
export enum TextAreaTargets {
  root = 'root',
  characterLimit = 'characterLimit',
  errors = 'errors',
  helperText = 'helperText',
  input = 'input',
  inputPrefix = 'inputPrefix',
  inputSuffix = 'inputSuffix',
  label = 'label',
  readOnly = 'readOnly',
  tooltipButton = 'tooltipButton',
  tooltipPopover = 'tooltipPopover',
}